+++ src/main/java/com/sprint/mission/discodeit/security/handler/CustomLoginSuccessHandler.java
@@ -1,4 +1,4 @@
-package com.sprint.mission.discodeit.common.security.handler;
+package com.sprint.mission.discodeit.security.handler;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.sprint.mission.discodeit.entity.CustomUserDetails;
